cleaning up some character stuff
